+++ pytorch_segmentation_models_trainer/tools/data_handlers/vector_reader.py
@@ -19,12 +19,15 @@
  ****
 """
 import abc
+import concurrent.futures
 import functools
+import itertools
 import operator
 import os
 from collections import defaultdict
 from dataclasses import MISSING, dataclass, field
 from enum import Enum
+from multiprocessing import Manager
 from pathlib import Path
 from typing import List, Union
 
@@ -36,8 +39,11 @@ from affine import Affine
 from bidict import bidict
 from geopandas import GeoDataFrame, GeoSeries
 from pycocotools.coco import COCO
+from pytorch_segmentation_models_trainer.tools.parallel_processing.process_executor import \
+    Executor
 from shapely.geometry import (GeometryCollection, LineString, MultiLineString,
                               MultiPoint, MultiPolygon, Point, Polygon, base)
+from tqdm import tqdm
 
 
 class GeomType(Enum):
@@ -50,6 +56,13 @@ suffix_dict = bidict({
     "GeoJSON": ".geojson",
     "ESRI Shapefile": ".shp"
 })
+
+def get_chunks(iterable, size, filler=None):
+    it = itertools.chain(iterable, itertools.repeat(filler,size-1))
+    chunk = tuple(itertools.islice(it,size))
+    while len(chunk) == size:
+        yield chunk
+        chunk = tuple(itertools.islice(it,size))
 
 @dataclass
 class GeoDF(abc.ABC):
@@ -169,32 +182,62 @@ class COCOMemoryGeoDF(GeoDF):
 @dataclass
 class COCOGeoDF:
     file_name: str = MISSING
+    max_workers: int = os.cpu_count()
+    chunk_size: int = 1000
+    pre_build_vector_dict: bool = True
 
     def __post_init__(self):
         self.coco = COCO(self.file_name)
-        self.vector_dict = {
-            image_id: COCOMemoryGeoDF(
-                map(
-                    _build_polygon_from_annotation,
-                    self.coco.loadAnns(ids=self.coco.getAnnIds(imgIds=image_id))
-                )
-            ) for image_id in self.coco.getImgIds(catIds=self.coco.getCatIds())
-        }
+        self.image_id_list = [id for id in self.coco.getImgIds(catIds=self.coco.getCatIds())]
+        self.vector_dict = dict()
+        if not self.pre_build_vector_dict:
+            return
+        def build_coco_memory_geodf_item(image_id_chunk):
+            return {
+                image_id: self.build_item(image_id) for image_id in image_id_chunk if image_id is not None
+            }
+        with concurrent.futures.ThreadPoolExecutor(max_workers=self.max_workers) as executor:
+            futures = set(
+                executor.submit(build_coco_memory_geodf_item, image_id_chunk) \
+                    for image_id_chunk in get_chunks(self.image_id_list, self.chunk_size)
+            )
+            kwargs = {
+                'total': len(futures),
+                'unit': 'chunks',
+                'unit_scale': True,
+                'leave': True,
+                'desc': "Building geodf from coco dataset"
+            }
+            for r in tqdm(concurrent.futures.as_completed(futures), **kwargs):
+                self.vector_dict.update(r.result())
+    
+    def build_item(self, image_id):
+        return COCOMemoryGeoDF(
+            map(
+                self._build_polygon_from_annotation,
+                self.coco.loadAnns(ids=self.coco.getAnnIds(imgIds=image_id))
+            )
+        )
     
     def get_geodf_item(self, key: str) -> GeoDataFrame:
-        return self.vector_dict[int(key)]
+        int_key = int(key)
+        if int_key not in self.image_id_list:
+            raise KeyError("key not in image list")
+        if int_key not in self.vector_dict:
+            self.vector_dict[int_key] = self.build_item(int_key)
+        return self.vector_dict[int_key]
 
 
-def _build_polygon_from_annotation(annotation):
-    seg_list = annotation["segmentation"]
-    if len(seg_list) != 1:
-        print(f"len(seg_list) = {len(seg_list)}, but it should be 1.")
-        raise NotImplementedError
-    coords = np.reshape(
-        np.array(seg_list),
-        (-1, 2)
-    )
-    return Polygon(coords)
+    def _build_polygon_from_annotation(self, annotation):
+        seg_list = annotation["segmentation"]
+        if len(seg_list) != 1:
+            print(f"len(seg_list) = {len(seg_list)}, but it should be 1.")
+            raise NotImplementedError
+        coords = np.reshape(
+            np.array(seg_list),
+            (-1, 2)
+        )
+        return Polygon(coords)
 
 def handle_features(input_features, output_type: GeomType = None, return_list: bool = False) -> list:
     if output_type is None: