+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 0.8.1
+
+- Minor improvements and bug fixes on polygon building inference;
+- Bug fixes on mask builder;
+- Performance improvement on mask builder using coco format;
+
 # Version 0.8.0
 
 - Added inference features;